requirements.txt
fastapi
uvicorn
supabase
cryptography
lightgbm
shap
joblib
pandas
numpy
python-dotenv
dice-ml
scikit-learn
